Mapper refactors
