+++ src/main/java/com/cyberfreak/services/mapper/ApplicationMapper.java
@@ -1,7 +1,7 @@
 package com.cyberfreak.services.mapper;
 
-import com.cyberfreak.services.api.request.CreateApplicationRequest;
-import com.cyberfreak.services.api.request.UpdateApplicationRequest;
+import com.cyberfreak.services.api.request.application.CreateApplicationRequest;
+import com.cyberfreak.services.api.request.application.UpdateApplicationRequest;
 import com.cyberfreak.services.domain.Application;
 import com.cyberfreak.services.dto.ApplicationDto;
 import org.mapstruct.*;
