+++ src/main/java/com/cyberfreak/services/mapper/ContentItemMapper.java
@@ -1,6 +1,6 @@
 package com.cyberfreak.services.mapper;
 
-import com.cyberfreak.services.api.request.CreateContentItemRequest;
+import com.cyberfreak.services.api.request.pagecontent.PageContentContentItemRequest;
 import com.cyberfreak.services.domain.ContentItem;
 import com.cyberfreak.services.dto.ContentItemDto;
 import com.cyberfreak.services.service.ApplicationService;
@@ -20,9 +20,7 @@ public interface ContentItemMapper extends MapperBase<ContentItem, ContentItemDt
     @BeanMapping(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE)
     ContentItem partialUpdate(ContentItemDto contentItemDto, @MappingTarget ContentItem contentItem);
 
-    @Named("mapCreateContentItemRequestToContentItemDtoWithContext")
     @Mapping(source = "key", target = "resourceMap.resourceKey")
     @Mapping(source = "value", target = "resourceMap.resourceValue")
-    @Mapping(source = "applicationId", target = "application", qualifiedByName = "mapApplicationIdToApplicationDtoWithContext")
-    ContentItemDto toDto(CreateContentItemRequest createContentItemRequest, @Context Long applicationId);
+    ContentItemDto toDto(PageContentContentItemRequest pageContentContentItemRequest);
 }
